+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Removed
+
+- cosmwasm-std: Remove `Context` - deprecated in previous release
+- cosmwasm-std: Remove `HandleResponse`, `InitResponse`, and `MigrateResponse` -
+  deprecated in previous release
+
 ### Changed
 
 - cosmwasm-std: Rename the `send` function parameter to `funds` in `WasmMsg`